+++ GameServer/packets/Server/PacketLib168.cs
@@ -774,18 +774,18 @@ namespace DOL.GS.PacketHandler
 				if (m_gameClient.Account.PrivLevel < 2)
 				{
 					// no name only if normal player
-					if ((npc.Flags & (uint) GameNPC.eFlags.CANTTARGET) != 0)
+					if ((npc.Flags & GameNPC.eFlags.CANTTARGET) != 0)
 						flags |= 0x01;
-					if ((npc.Flags & (uint) GameNPC.eFlags.DONTSHOWNAME) != 0)
+					if ((npc.Flags & GameNPC.eFlags.DONTSHOWNAME) != 0)
 						flags |= 0x02;
 				}
-				if ((npc.Flags & (uint) GameNPC.eFlags.STATUE) != 0)
+				if ((npc.Flags & GameNPC.eFlags.STATUE) != 0)
 				{
 					flags |= 0x01;
 				}
 				if (npc.IsUnderwater)
 					flags |= 0x10;
-				if ((npc.Flags & (uint) GameNPC.eFlags.FLYING) != 0)
+				if ((npc.Flags & GameNPC.eFlags.FLYING) != 0)
 					flags |= 0x20;
 
 				if (npc.IsMoving && !npc.IsAtTargetPosition)
@@ -1016,11 +1016,11 @@ namespace DOL.GS.PacketHandler
 				pak.WriteByte(npc.GetDisplayLevel(m_gameClient.Player));
 
 				var flags = (byte) (GameServer.ServerRules.GetLivingRealm(m_gameClient.Player, npc) << 6);
-				if ((npc.Flags & (uint) GameNPC.eFlags.GHOST) != 0) flags |= 0x01;
+				if ((npc.Flags & GameNPC.eFlags.GHOST) != 0) flags |= 0x01;
 				if (npc.Inventory != null)
 					flags |= 0x02; //If mob has equipment, then only show it after the client gets the 0xBD packet
-				if ((npc.Flags & (uint) GameNPC.eFlags.PEACE) != 0) flags |= 0x10;
-				if ((npc.Flags & (uint) GameNPC.eFlags.FLYING) != 0) flags |= 0x20;
+				if ((npc.Flags & GameNPC.eFlags.PEACE) != 0) flags |= 0x10;
+				if ((npc.Flags & GameNPC.eFlags.FLYING) != 0) flags |= 0x20;
 
 				pak.WriteByte(flags);
 				pak.WriteByte(0x20); //TODO this is the default maxstick distance
@@ -1028,9 +1028,9 @@ namespace DOL.GS.PacketHandler
 				string add = "";
 				if (m_gameClient.Account.PrivLevel > 1)
 				{
-					if ((npc.Flags & (uint) GameNPC.eFlags.CANTTARGET) != 0)
+					if ((npc.Flags & GameNPC.eFlags.CANTTARGET) != 0)
 						add += "-DOR"; // indicates DOR flag for GMs
-					if ((npc.Flags & (uint) GameNPC.eFlags.DONTSHOWNAME) != 0)
+					if ((npc.Flags & GameNPC.eFlags.DONTSHOWNAME) != 0)
 						add += "-NON"; // indicates NON flag for GMs
 				}
 				string name = npc.Name;
